+++ example_project/lib/example_web/live/live_chat.ex
@@ -21,7 +21,7 @@ defmodule ExampleWeb.LiveExample6 do
           <button class="py-2 px-4 bg-black text-white rounded">Join</button>
         </form>
       <% else %>
-        <LiveSvelte.render
+        <.svelte
           name="Chat"
           props={%{messages: @messages, name: @name}}
           class="w-full h-full flex justify-center items-center"