+++ examples/log_list/live_svelte.ex
@@ -3,7 +3,7 @@ defmodule ExamplesWeb.LogListLive do
 
   def render(assigns) do
     ~H"""
-    <LiveSvelte.render name="LogList" props={%{items: @items}} />
+    <LiveSvelte.svelte name="LogList" props={%{items: @items}} />
     """
   end