+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2023-05-26
+
+### Added
+
+-  Install instructions on importing LiveSvelte inside `html_helpers` so `<.svelte />` is possible.
+
+### Changed
+
+-  Deprecated `LiveSvelte.render` in favor of `LiveSvelte.svelte`
+-  Now we're using `<.svelte />` instead of `<LiveSvelte.render />` in the examples
+
 ## [0.5.1] - 2023-05-06
 
 ### Fixed